Refactoring
Put workflow methods dealing with versions and branches in a separate
class. Use lazy properties to short down the execution
time. snpseq-paths field, workflow is replaced with branch-provider,
which is an instance of the new class.

Lazy properties are implemented on latest version (i.e. current
operation version), and candidate branch.